requirements.txt
# VitFly-AirSim Requirements for Windows
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0

# AirSim and simulation
airsim>=1.8.1
msgpack-rpc-python>=0.4.1

# Data processing and visualization
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scipy>=1.7.0

# Configuration and logging
PyYAML>=5.4.0
tensorboard>=2.7.0
tqdm>=4.62.0
colorama>=0.4.4

# Image processing
scikit-image>=0.18.0

# Windows specific
pywin32>=227; sys_platform == "win32"

# Optional: For GPU monitoring
pynvml>=11.0.0

# Development and testing
pytest>=6.2.0
black>=21.9.0
isort>=5.9.0